requirements.txt
requests
mysql-connector-python
schedule
twilio
python-dotenv
